+++ src/api/index.js
@@ -1,11 +1,10 @@
 const express = require("express");
 
-const { healthRouter } = require('../routes/health/health.router');
-const cohortRouter = require('../routes/cohorts');
-
+const { healthRouter } = require("../routes/health/health.router");
+const { cohortsRouter } = require("../routes/cohorts/cohorts.router");
 
 const router = express.Router();
 router.use("/health", healthRouter);
-router.use('/cohorts', cohortRouter);
+router.use("/cohorts", cohortsRouter);
 
 module.exports = router;
